SoLU softmax needs to be computed in float32

+++ src/models/layers.py
@@ -40,7 +40,7 @@ def create_mask(seq_len_k, slopes):
 
 
 def solu(x: jnp.array) -> jnp.array:
-    return x * jax.nn.softmax(x, axis=-1)
+    return x * jax.nn.softmax(x.astype(jnp.float32), axis=-1)
 
 
 class MLPBlock(nn.Module):
